+++ jwst/datamodels/tests/test_wcs.py
@@ -4,7 +4,10 @@ import warnings
 import numpy as np
 from numpy.testing import assert_array_almost_equal
 
+from ..wcs_ref_models import FilteroffsetModel
 from .. import ImageModel
+
+import pytest
 
 
 FITS_FILE = os.path.join(os.path.dirname(__file__), 'data', 'sip.fits')
@@ -57,3 +60,37 @@ def test_wcs(tmpdir):
         wcs5 = dm5.get_fits_wcs()
 
     assert wcs5.wcs.crpix[0] == 42.0
+
+
+def test_wcs_ref_models():
+    filters = [{'name': 'F090W', 'row_offset': 1, 'column_offset': 1},
+               {'name': 'F070W', 'row_offset': 2, 'column_offset': 2}
+              ]
+    with FilteroffsetModel(filters=filters, instrument='NIRCAM', strict_validation=True) as fo:
+        fo.filters == filters
+        with pytest.raises(ValueError) as e:
+            fo.validate()
+            assert str(e.value()) == ("Model.meta is missing values for "
+                                     "['description', 'reftype', 'author', 'pedigree',"
+                                     "'useafter']")
+
+    filters = [{'filter': 'F090W', 'pupil': 'GRISMR',
+                'row_offset': 1, 'column_offset': 1},
+               {'filter': 'F070W', 'pupil': 'GRISMC',
+                'row_offset': 2, 'column_offset': 2}
+              ]
+    with FilteroffsetModel(filters=filters, instrument='NIRCAM', strict_validation=True) as fo:
+        fo.filters == filters
+        fo.meta.description = "Filter offsets"
+        fo.meta.reftype = "filteroffset"
+        fo.meta.author = "Unknown"
+        fo.meta.pedigree = "GROUND"
+        fo.meta.useafter = "2019-12-01"
+
+        with pytest.raises(ValueError) as e:
+            fo.validate()
+            assert str(e.value()) == ("Expected meta.instrument.channel for instrument NIRCAM to be one of "
+                                      "['SHORT', 'LONG']")
+        fo.meta.instrument.channel = 'SHORT'
+        fo.meta.instrument.module="A"
+        fo.validate()